Rename Bot.py to Tg_bot_send_message.py
Here's a script for a Telegram bot which will send messages. Just enter the number of messages to send and the initial sending speed in milliseconds. There is such thing as telegram's flood security limitations, so here is a 'fuse' mechanism that slow down the sending speed. So if telegram activates its flood security feature, the script will continue to run, but message sending speed will be reduced. I can't break it's limitations with my knowledge, but it isn't for long

To use the script, you'll need to enter your Telegram bot token and the ID of the chat where you want the messages to be sent. You can obtain the chat ID without administrative rights by opening the chat in Chrome and finding the ID in the last symbols of the URL. After you add the bot, you can start the script from any editor. Alternatively, you can insert the command into the bot, which will eliminate the need to enter the bot token.